requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
langgraph>=0.2.0
langchain-core>=0.2.27
langchain-openrouter>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
httpx>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
